use CurrentUICulture to get culture use Path.GetFileName for safer filename parsing of the url split Startup classes based on feature to reduce usings
