+++ src/OrchardCore.Modules/OrchardCore.Facebook/ScriptsMiddleware.cs
@@ -1,3 +1,5 @@
+using System.Globalization;
+using System.IO;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -21,14 +23,14 @@ namespace OrchardCore.Facebook
 
         public async Task Invoke(HttpContext httpContext)
         {
-            string script = null;
             if (httpContext.Request.Path.StartsWithSegments("/OrchardCore.Facebook/sdk"))
             {
+                var script = default(string);
                 var site = (await _siteService.GetSiteSettingsAsync());
                 var settings = site.As<FacebookSettings>();
-                if (httpContext.Request.Path.Value.EndsWith("fbsdk.js"))
+                if (Path.GetFileName(httpContext.Request.Path.Value) == "fbsdk.js")
                 {
-                    var locale = string.IsNullOrWhiteSpace(site.Culture) ? "en_US" : site.Culture.Replace("-", "_");
+                    var locale = CultureInfo.CurrentUICulture.Name.Replace("-", "_");
                     script = $@"(function(d){{
                         var js, id = 'facebook-jssdk'; if (d.getElementById(id)) {{ return; }}
                         js = d.createElement('script'); js.id = id; js.async = true;
@@ -36,7 +38,7 @@ namespace OrchardCore.Facebook
                         d.getElementsByTagName('head')[0].appendChild(js);
                     }} (document));";
                 }
-                if (httpContext.Request.Path.Value.EndsWith("fb.js"))
+                else if (Path.GetFileName(httpContext.Request.Path.Value) == "fb.js")
                 {
                     if (!string.IsNullOrWhiteSpace(settings?.AppId))
                     {
@@ -52,18 +54,17 @@ namespace OrchardCore.Facebook
                         script = $"window.fbAsyncInit = function(){{ FB.init({options});}};";
                     }
                 }
+
+                if (script != null)
+                {
+                    var bytes = Encoding.UTF8.GetBytes(script);
+                    var cancellationToken = httpContext?.RequestAborted ?? CancellationToken.None;
+                    await httpContext.Response.Body.WriteAsync(Encoding.UTF8.GetBytes(script), 0, bytes.Length, cancellationToken);
+                    return;
+                }
             }
 
-            if (script == null)
-            {
-                await _next.Invoke(httpContext);
-            }
-            else
-            {
-                var bytes = Encoding.UTF8.GetBytes(script);
-                var cancellationToken = httpContext?.RequestAborted ?? CancellationToken.None;
-                await httpContext.Response.Body.WriteAsync(Encoding.UTF8.GetBytes(script), 0, bytes.Length, cancellationToken);
-            }
+            await _next.Invoke(httpContext);
         }
     }
 }
