+++ src/OrchardCore.Modules/OrchardCore.Facebook/Startup.cs
@@ -1,37 +1,17 @@
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Facebook;
+using System;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Options;
-using OrchardCore.ContentManagement;
-using OrchardCore.ContentManagement.Handlers;
-using OrchardCore.ContentManagement.Display.ContentDisplay;
-using OrchardCore.ContentTypes.Editors;
-using OrchardCore.Data.Migration;
 using OrchardCore.DisplayManagement.Handlers;
 using OrchardCore.Facebook.Drivers;
 using OrchardCore.Facebook.Filters;
-using OrchardCore.Facebook.Login.Configuration;
-using OrchardCore.Facebook.Login.Drivers;
-using OrchardCore.Facebook.Login.Services;
 using OrchardCore.Facebook.Services;
-using OrchardCore.Facebook.Widgets;
-using OrchardCore.Facebook.Widgets.Drivers;
-using OrchardCore.Facebook.Widgets.Handlers;
-using OrchardCore.Facebook.Widgets.Models;
-using OrchardCore.Facebook.Widgets.Settings;
 using OrchardCore.Modules;
 using OrchardCore.Navigation;
+using OrchardCore.ResourceManagement;
 using OrchardCore.Security.Permissions;
 using OrchardCore.Settings;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Routing;
-using System;
-using OrchardCore.ResourceManagement;
-using ResourceManifest = OrchardCore.Facebook.ResourceManifest;
-using OrchardCore.DisplayManagement.Descriptors;
-using OrchardCore.Facebook.Widgets.Services;
 
 namespace OrchardCore.Facebook
 {
@@ -59,41 +39,4 @@ namespace OrchardCore.Facebook
 
         }
     }
-
-    [Feature(FacebookConstants.Features.Login)]
-    public class LoginStartup : StartupBase
-    {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddSingleton<IFacebookLoginService, FacebookLoginService>();
-            services.AddScoped<IDisplayDriver<ISite>, FacebookLoginSettingsDisplayDriver>();
-            services.AddScoped<INavigationProvider, AdminMenuLogin>();
-
-            // Register the options initializers required by the OpenID Connect client handler.
-            services.TryAddEnumerable(new[]
-            {
-                // Orchard-specific initializers:
-                ServiceDescriptor.Transient<IConfigureOptions<AuthenticationOptions>, FacebookLoginConfiguration>(),
-                ServiceDescriptor.Transient<IConfigureOptions<FacebookOptions>, FacebookLoginConfiguration>(),
-
-                // Built-in initializers:
-                ServiceDescriptor.Transient<IPostConfigureOptions<FacebookOptions>, OAuthPostConfigureOptions<FacebookOptions,FacebookHandler>>()
-            });
-        }
-    }
-
-    [Feature(FacebookConstants.Features.Widgets)]
-    public class WidgetsStartup : StartupBase
-    {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddScoped<IContentPartDisplayDriver, FacebookPluginPartDisplayDriver>();
-            services.AddScoped<IShapeTableProvider, LiquidShapes>();
-            services.AddSingleton<ContentPart, FacebookPluginPart>();
-            services.AddScoped<IDataMigration, WidgetMigrations>();
-            services.AddScoped<IContentPartHandler, FacebookPluginPartHandler>();
-            services.AddScoped<IContentTypePartDefinitionDisplayDriver, FacebookPluginPartSettingsDisplayDriver>();
-        }
-    }
-
 }
